breakup compose/project.go a bit
