+++ internal/providers/docker/compose/config.go
@@ -1,0 +1,7 @@
+package compose
+
+type Config struct {
+	File     string `yaml:"file,omitempty"`
+	External bool   `yaml:"external,omitempty"`
+	Name     string `yaml:"name,omitempty"`
+}
